Port WIP
Build the model project from Java in C#.  Also implemented some of the extra code after the XSD generated code.  It looks to support a tools project and is parsing Java Language.  I tapped the brakes and will come back to this.  I am new to Antlr, ELM, CQL and all of this.  Anyway want to safe my work and start again.